pyproject.toml
[tool.poetry]
name = "olive-shop"
version = "0.1.0"
description = ""
authors = ["dmitry <kharisov.dmitry101@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
environs = "^11.0.0"
psycopg = "^3.1.19"
sqlalchemy = "^2.0.30"
sqladmin = "^0.17.0"
fastapi-storages = {extras = ["full"], version = "^0.3.0"}
pyjwt = "^2.8.0"
asyncpg = "^0.29.0"
itsdangerous = "^2.2.0"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
